move RHESSI loader to extern
