+++ sunkit_spex/extern/__init__.py
@@ -5,3 +5,11 @@ are external to it, and hence are developed in a separate source tree. Note
 that this package is distinct from the /cextern directory of the source code
 distribution, as that directory only contains C extension code.
 """
+
+from . import (
+    rhessi
+)
+
+__all__ = [
+    'rhessi',
+]
