+++ sunkit_spex/fitting_legacy/__init__.py
@@ -8,6 +8,4 @@ from . import (
     parameter_handler,
     photon_models_for_fitting,
     rainbow_text,
-    rhes_spec_code,
-    stix_spec_code,
 )
